Create UVa12403- Save Setu
Tips: a) If a is a character array and we input it just like a string, then the input process will not quit until it meets a stop character. (Enter, Blank and   Table).;
b) The use of strcmp defined in <string.h>. This function can compare two string, if they are same, then return zero. If not, return 1.