Create UVa12250-Language Detection
Tips: a) We can store up all the finite possible output in one or more array(s) and then to look up it.
b) The use of bool: It is aimed for the search situations. If we found the ideal output, then found can be true, If not, then bool will be false.
c) The input form of string array. It should print out like this: printf("%s", stringarray[i].c_str()). Pay attention to not left out the suffix.